JD-7 backend

+++ src/main/java/ru/mail/jira/plugins/groovy/api/dto/audit/AuditLogEntryForm.java
@@ -17,7 +17,7 @@ public class AuditLogEntryForm {
     @XmlElement
     private AuditCategory category;
     @XmlElement
-    private Long entityId;
+    private Integer entityId;
     @XmlElement
     private AuditAction action;
     @XmlElement
